Modularized files
Used, I beleive, ES6, modularization to break out functions
into different files.
