+++ modules/data-models.js
@@ -1,0 +1,4 @@
+//constructors
+
+exports.Location = Location;
+exports.WeatherDay = WeatherDay;
