+++ modules/invalid-path.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const invalidAPIPathError = response => {
+  response.status(404).send('Invalid API path.');
+}
+
+module.exports.invalidAPIPathError = invalidAPIPathError;
